backend/requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
flask-cors
waitress
python-dotenv
